Integrate custom trigger shortcut with default TAB
The integration made by A. Bond allows to change the trigger to
any valid keyboard shortcut but the only way to restore to TAB
was removing the parameter that contains the new shortcut.
The reason? When you press TAB with the intention to make it the
shortcut instead of being taken by the QKeySequenceEdit the GUI
interpreted the TAB key press as a command to pass to next element.

This commit adds a checkbox to let the user override the default TAB
shortcut, after changing the shortcut you can came back to TAB uncheking
the checkbox.

Fix #3